+++ packages/superdeck/test/src/builtins/webview_widget_test.dart
@@ -699,7 +699,7 @@ class _TwoWebViewsHarness extends StatelessWidget {
   Widget build(BuildContext context) {
     final slide = SlideConfiguration(
       slideIndex: 0,
-      style: SlideStyle(),
+      style: SlideStyler(),
       slide: Slide(
         key: 'slide-0',
         sections: [
@@ -756,7 +756,7 @@ class _WebViewHarness extends StatelessWidget {
   Widget build(BuildContext context) {
     final slide = SlideConfiguration(
       slideIndex: 0,
-      style: SlideStyle(),
+      style: SlideStyler(),
       slide: Slide(
         key: 'slide-0',
         sections: [