+++ packages/superdeck/test/src/markdown/markdown_builders_test.dart
@@ -237,7 +237,7 @@ class _MarkdownHarness extends StatelessWidget {
     final styleSheet = slideSpec.toStyle();
     final slideConfiguration = SlideConfiguration(
       slideIndex: 0,
-      style: SlideStyle(),
+      style: SlideStyler(),
       slide: Slide(key: 'slide'),
       thumbnailKey: 'thumb.png',
     );